Massive cleanup of obsolete artifacts after removal of dependency on ICriteria.
