+++ Application/EdFi.Ods.Common/Providers/Criteria/PagedAggregateIdsCriteriaProvider.cs
@@ -5,6 +5,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.Linq;
 using EdFi.Common.Configuration;
 using EdFi.Ods.Common.Configuration;
@@ -12,11 +13,12 @@ using EdFi.Ods.Common.Database.Querying;
 using EdFi.Ods.Common.Database.Querying.Dialects;
 using EdFi.Ods.Common.Descriptors;
 using EdFi.Ods.Common.Extensions;
+using EdFi.Ods.Common.Infrastructure.Repositories;
 using EdFi.Ods.Common.Models;
 using EdFi.Ods.Common.Models.Domain;
+using EdFi.Ods.Common.Models.Queries;
 using EdFi.Ods.Common.Specifications;
 using NHibernate;
-using NHibernate.Persister.Entity;
 
 namespace EdFi.Ods.Common.Providers.Criteria
 {
@@ -24,12 +26,15 @@ namespace EdFi.Ods.Common.Providers.Criteria
     /// Builds a query that retrieves the Ids for the next page of data.
     /// </summary>
     /// <typeparam name="TEntity">The type of entity for which to build the query.</typeparam>
-    public class PagedAggregateIdsCriteriaProvider<TEntity> : AggregateRootCriteriaProviderBase<TEntity>, IPagedAggregateIdsCriteriaProvider<TEntity>
+    public class PagedAggregateIdsCriteriaProvider<TEntity> : NHibernateRepositoryOperationBase, IPagedAggregateIdsCriteriaProvider<TEntity>
         where TEntity : class
     {
+        private readonly IDescriptorResolver _descriptorResolver;
+        private readonly IPersonEntitySpecification _personEntitySpecification;
         private readonly IDomainModelProvider _domainModelProvider;
         private readonly Dialect _dialect;
         private readonly DatabaseEngine _databaseEngine;
+        private readonly IPersonTypesProvider _personTypesProvider;
 
         public PagedAggregateIdsCriteriaProvider(
             ISessionFactory sessionFactory, 
@@ -40,77 +45,26 @@ namespace EdFi.Ods.Common.Providers.Criteria
             IDomainModelProvider domainModelProvider,
             Dialect dialect,
             DatabaseEngine databaseEngine)
-            : base(sessionFactory, descriptorResolver, personEntitySpecification, personTypesProvider)
+            : base(sessionFactory)
         {
+            _descriptorResolver = descriptorResolver;
+            _personEntitySpecification = personEntitySpecification;
             _domainModelProvider = domainModelProvider;
             _dialect = dialect;
             _databaseEngine = databaseEngine;
-
-            _identifierColumnNames = new Lazy<string[]>(
-                () =>
-                {
-                    var persister = (AbstractEntityPersister) SessionFactory.GetClassMetadata(typeof(TEntity));
-
-                    if (persister.IdentifierColumnNames != null && persister.IdentifierColumnNames.Length > 0)
-                    {
-                        return persister.IdentifierColumnNames;
-                    }
-
-                    return new[] { "Id" };
-                });
+            _personTypesProvider = personTypesProvider;
 
             _defaultPageLimitSize = defaultPageSizeLimitProvider.GetDefaultPageSizeLimit();
         }
 
-        private readonly Lazy<string[]> _identifierColumnNames;
         private readonly int _defaultPageLimitSize;
 
-        // /// <summary>
-        // /// Get a <see cref="NHibernate.ICriteria"/> query that retrieves the Ids for the next page of data.
-        // /// </summary>
-        // /// <param name="specification">An instance of the entity containing parameters to be added to the query.</param>
-        // /// <param name="queryParameters">The query parameters to be applied to the filtering.</param>
-        // /// <returns>The NHibernate <see cref="NHibernate.ICriteria"/> instance representing the query.</returns>
-        // public ICriteria GetCriteriaQuery(TEntity specification, IQueryParameters queryParameters)
-        // {
-        //     var idQueryCriteria = Session.CreateCriteria<TEntity>("aggregateRoot")
-        //         .SetProjection(Projections.Distinct(GetColumnProjectionsForDistinctWithOrderBy()))
-        //         .SetFirstResult(queryParameters.Offset ?? 0)
-        //         .SetMaxResults(queryParameters.Limit ?? _defaultPageLimitSize);
-        //
-        //     AddDefaultOrdering(idQueryCriteria);
-        //
-        //     // Add specification-based criteria
-        //     ProcessSpecification(idQueryCriteria, specification);
-        //
-        //     // Add special query fields
-        //     ProcessQueryParameters(idQueryCriteria, queryParameters);
-        //
-        //     return idQueryCriteria;
-        //     
-        //     IProjection GetColumnProjectionsForDistinctWithOrderBy()
-        //     {
-        //         var projections = Projections.ProjectionList();
-        //     
-        //         // Add the resource identifier (this is the value we need for the secondary "page" query)
-        //         projections.Add(Projections.Property("Id"));
-        //     
-        //         // Add the order by (primary key) columns (required when using DISTINCT with ORDER BY)
-        //         foreach (var identifierColumnName in _identifierColumnNames.Value)
-        //         {
-        //             projections.Add(Projections.Property(identifierColumnName));
-        //         }
-        //
-        //         return projections;
-        //     }
-        // }
-
         /// <summary>
-        /// Get a <see cref="NHibernate.ICriteria"/> query that retrieves the Ids for the next page of data.
+        /// Get a <see cref="QueryBuilder"/> containing the query that retrieves the Ids for the next page of data.
         /// </summary>
         /// <param name="specification">An instance of the entity containing parameters to be added to the query.</param>
         /// <param name="queryParameters">The query parameters to be applied to the filtering.</param>
-        /// <returns>The NHibernate <see cref="NHibernate.ICriteria"/> instance representing the query.</returns>
+        /// <returns>The <see cref="QueryBuilder"/> instance representing the query.</returns>
         public QueryBuilder GetQueryBuilder(TEntity specification, IQueryParameters queryParameters)
         {
             var idQueryBuilder = new QueryBuilder(_dialect);
@@ -122,7 +76,7 @@ namespace EdFi.Ods.Common.Providers.Criteria
                 throw new Exception($"Unable to find API model entity for '{entityFullName}'.");
             }
 
-            // TODO: Need physical table name here -- entity.TableName()
+            // TODO: ODS-6444 - Need physical table name here (and everywhere, really) -- entity.TableName()
             var tableName = entity.FullName.ToString();
 
             string[] selectColumns = GetColumnProjectionsForDistinctWithOrderBy(entity).ToArray();
@@ -132,7 +86,7 @@ namespace EdFi.Ods.Common.Providers.Criteria
                 .Select(selectColumns)
                 .LimitOffset(queryParameters.Limit ?? _defaultPageLimitSize, queryParameters.Offset ?? 0);
 
-            // TODO: In order for query caching to work, limit/offset must be parameterized in query (not embedded as literal values)
+            // TODO: ODS-6444 - In order for query caching to work, limit/offset must be parameterized in query (not embedded as literal values)
 
             // Add the join to the base type
             if (entity.IsDerived)
@@ -154,7 +108,7 @@ namespace EdFi.Ods.Common.Providers.Criteria
 
             AddDefaultOrdering(idQueryBuilder, entity);
 
-            // TODO: Consider cloning the querybuilder at this point and introducing a seam for applying the additional parameters so that authorization strategies can also be incorporated and cloned
+            // TODO: ODS-6444 - Consider cloning the querybuilder at this point and introducing a seam for applying the additional parameters so that authorization strategies can also be incorporated and cloned
 
             // Add specification-based criteria
             ProcessSpecification(idQueryBuilder, specification, entity);
@@ -170,10 +124,9 @@ namespace EdFi.Ods.Common.Providers.Criteria
                 yield return "Id";
             
                 // Add the order by (primary key) columns (required when using DISTINCT with ORDER BY)
-                // foreach (var identifierColumnName in _identifierColumnNames.Value)
                 foreach (var identifierProperty in (entity.BaseEntity ?? entity).Identifier.Properties)
                 {
-                    string identifierColumnName = identifierProperty.ColumnName(_databaseEngine, "XYZ");
+                    string identifierColumnName = identifierProperty.ColumnName(_databaseEngine, identifierProperty.PropertyName);
                     
                     if (entity.IsDerived)
                     {
@@ -185,30 +138,141 @@ namespace EdFi.Ods.Common.Providers.Criteria
                     }
                 }
             }
+
+            void AddDefaultOrdering(QueryBuilder queryBuilder, Entity entity)
+            {
+                foreach (var identifierProperty in (entity.BaseEntity ?? entity).Identifier.Properties)
+                {
+                    string identifierColumnName = identifierProperty.ColumnName(_databaseEngine, identifierProperty.PropertyName);
+
+                    if (entity.IsDerived)
+                    {
+                        queryBuilder.OrderBy($"b.{identifierColumnName}");
+                    }
+                    else
+                    {
+                        queryBuilder.OrderBy(identifierColumnName);
+                    }
+                }
+            }
+            
+            static void ProcessQueryParameters(QueryBuilder queryBuilder, IQueryParameters parameters)
+            {
+                foreach (IQueryCriteriaBase criteria in parameters.QueryCriteria)
+                {
+                    if (criteria is TextCriteria textCriteria)
+                    {
+                        MatchMode mode;
+
+                        switch (textCriteria.MatchMode)
+                        {
+                            case TextMatchMode.Anywhere:
+                                mode = MatchMode.Anywhere;
+
+                                break;
+
+                            case TextMatchMode.Start:
+                                mode = MatchMode.Start;
+
+                                break;
+
+                            case TextMatchMode.End:
+                                mode = MatchMode.End;
+
+                                break;
+
+                            //case TextMatchMode.Exact:
+                            default:
+                                mode = MatchMode.Exact;
+
+                                break;
+                        }
+
+                        queryBuilder.WhereLike(textCriteria.PropertyName, textCriteria.Value, mode);
+                    }
+                }
+            }
         }
 
-        // private void AddDefaultOrdering(ICriteria queryCriteria)
-        // {
-        //     foreach (var identifierColumnName in _identifierColumnNames.Value)
-        //     {
-        //         queryCriteria.AddOrder(Order.Asc(identifierColumnName));
-        //     }
-        // }
-
-        private void AddDefaultOrdering(QueryBuilder queryBuilder, Entity entity)
+        private void ProcessSpecification(QueryBuilder queryBuilder, TEntity specification, Entity entity)
         {
-            // foreach (var identifierColumnName in _identifierColumnNames.Value)
-            foreach (var identifierProperty in (entity.BaseEntity ?? entity).Identifier.Properties)
+            if (specification != null)
             {
-                string identifierColumnName = identifierProperty.ColumnName(_databaseEngine, "XYZ");
+                var propertyValuePairs = specification.ToDictionary(
+                    (descriptor, o) => ShouldIncludeInQueryCriteria(descriptor, o, specification));
 
-                if (entity.IsDerived)
+                foreach (var key in propertyValuePairs.Keys)
                 {
-                    queryBuilder.OrderBy($"b.{identifierColumnName}");
+                    IHasLookupColumnPropertyMap map = specification as IHasLookupColumnPropertyMap;
+
+                    if (map.IdPropertyByLookupProperty.TryGetValue(key, out LookupColumnDetails columnDetails))
+                    {
+                        // Look up the corresponding lookup id value from the cache
+                        var lookupId = _descriptorResolver.GetDescriptorId(
+                            columnDetails.LookupTypeName,
+                            Convert.ToString(propertyValuePairs[key]));
+
+                        // Add criteria for the lookup Id value, to avoid need to incorporate an INNER JOIN into the query
+                        if (propertyValuePairs[key] != null)
+                        {
+                            queryBuilder.Where(columnDetails.PropertyName, lookupId);
+                        }
+                        else
+                        {
+                            queryBuilder.WhereNull(key);
+                        }
+                    }
+                    else
+                    {
+                        string alias = (!entity.IsDerived || entity.PropertyByName.ContainsKey(key))
+                            ? "r"
+                            : "b";
+
+                        // Add the property equality condition to the query criteria
+                        if (propertyValuePairs[key] != null)
+                        {
+                            queryBuilder.Where($"{alias}.{key}", propertyValuePairs[key]);
+                        }
+                        else
+                        {
+                            queryBuilder.WhereNull($"{alias}.{key}");
+                        }
+                    }
                 }
-                else
+            }
+
+            bool ShouldIncludeInQueryCriteria(PropertyDescriptor property, object value, TEntity entity)
+            {
+                // Null values and underscore-prefixed properties are ignored for specification purposes
+                if (value == null || property.Name.StartsWith("_") || "|Url|".Contains((string)property.Name))
                 {
-                    queryBuilder.OrderBy(identifierColumnName);
+                    // TODO: Come up with better way to exclude non-data properties
+                    return false;
+                }
+
+                Type valueType = value.GetType();
+
+                // Only use value types (or strings), and non-default values (i.e. ignore 0's)
+                var result = (valueType.IsValueType || valueType == typeof(string))
+                    && (!value.Equals(valueType.GetDefaultValue())
+                        || (UniqueIdConventions.IsUSI(property.Name)
+                            && GetPropertyValue(entity, UniqueIdConventions.GetUniqueIdPropertyName(property.Name)) != null));
+
+                // Don't include properties that are explicitly to be ignored
+                result = result && !AggregateRootCriteriaProviderHelpers.PropertiesToIgnore.Contains(property.Name);
+
+                // Don't include UniqueId properties when they appear on a Person entity
+                result = result
+                    && (!AggregateRootCriteriaProviderHelpers.GetUniqueIdProperties(_personTypesProvider).Contains(property.Name)
+                        || _personEntitySpecification.IsPersonEntity(entity.GetType()));
+
+                return result;
+            
+                object GetPropertyValue(TEntity entity, string propertyName)
+                {
+                    var properties = entity.ToDictionary();
+
+                    return properties.Where(p => p.Key == propertyName).Select(p => p.Value).SingleOrDefault();
                 }
             }
         }
